rename CbasMockResponses -> cbas-sample-configurations
