+++ src/libs/ajax/workflows-app/cbas-sample-configurations.js
